+++ src/index.js
@@ -4723,9 +4723,17 @@ class ULabel {
     }
 
     swap_frame_image(new_src, frame=0) {
-        let ret = $(`img#${this.config["image_id_pfx"]}__${frame}`).attr("src");
+        const ret = $(`img#${this.config["image_id_pfx"]}__${frame}`).attr("src");
         $(`img#${this.config["image_id_pfx"]}__${frame}`).attr("src", new_src);
         return ret;
+    }
+
+    // Swap annotation box background color
+    swap_anno_bg_color(new_bg_color) {
+        const annbox = $("#" + this.config["annbox_id"]);
+        const ret = annbox.css("background-color");
+        annbox.css("background-color", new_bg_color);
+        return ret
     }
 
     // Change frame